refactor: 이벤트 참여 로직 클래스 분리
